test: extract shared input helper and template-literal clamp descriptions
`setZoom` in preview/index.test.tsx and `typeAndCommit` in
zoom-controls.test.tsx were near-duplicates of the same focus →
set-value → keydown sequence. Lift to `src/test/input.ts`.

Also: the clamp-test descriptions hardcoded "(10%)" / "(5000%)" while
deriving the actual assertions from MIN_SCALE / MAX_SCALE — if either
constant moves, the description drifts silently. Use template literals
so they stay in sync.
